Extract `wrap-method` tests from `sinon-test`.
Pulls out 3 suites into `test/util/core/wrap-method-test` for better organisation.
